+++ kmuhelper/emails/__init__.py
@@ -1,0 +1,1 @@
+from kmuhelper.emails.urls import urlpatterns