+++ kmuhelper/models.py
@@ -1,6 +1,5 @@
 # pylint: disable=unused-wildcard-import 
 
 from kmuhelper.main.models import *
-# Ansprechpartner, Bestellungskosten, Bestellungsposten, Bestellung, Kategorie, Kosten, Kunde, Lieferant, Lieferungsposten, Lieferung, Notiz, Produktkategorie, Produkt, Zahlungsempfaenger, Einstellung, Geheime_Einstellung
 from kmuhelper.app.models import *
-# ToDoNotiz, ToDoVersand, ToDoZahlungseingang, ToDoLagerbestand, ToDoLieferung
+from kmuhelper.emails.models import *