+++ kmuhelper/urls.py
@@ -5,6 +5,15 @@ from django.conf import settings
 from kmuhelper import views
 from kmuhelper.main.models import Einstellung, Geheime_Einstellung
 
+#####
+
+from rich import print
+
+prefix = "[deep_pink4][KMUHelper Urls][/] -"
+
+def log(string, *args):
+    print(prefix, string, *args)
+
 # Einstellungen festlegen
 
 try:
@@ -16,7 +25,7 @@ try:
     Geheime_Einstellung.objects.get_or_create(id="wc-consumer_secret")
     Geheime_Einstellung.objects.get_or_create(id="wc-url")
 except Exception as e:
-    print("[urls.py] - Error while adding settings:", e)
+    log("[urls.py] - Error while adding settings:", e)
 
 
 #######################
@@ -27,10 +36,11 @@ urlpatterns = [
     path('admin/', views.admin, name="admin"),
 
     path('_templatetest/<path:templatename>', views._templatetest, name="_templatetest"),
-    
+
     path('',        include('kmuhelper.main')),
     path('api/',    include('kmuhelper.api')),
     path('app/',    include('kmuhelper.app')),
+    path('emails/', include('kmuhelper.emails')),
     path('stats/',  include('kmuhelper.stats')),
     path('wc/',     include('kmuhelper.integrations.woocommerce')),
 ]